main.py
# Apply asynchronous processing (done)
# Apply incremental redis update (done)
# Apply CPU/memory usage limitation for scanning (done)
# Use backoff to check redis status for initial scanning (done)
# Fix some environment errors (done)
# Apply status health check api (done)
# Apply packets per protocol counting, using t-shark (done)
# Apply fuzzy searching (done)
# Apply configurable mounted directory

import redis
from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware
import asyncio
import backoff

from services.context import get_app_context
from services.logger import get_logger
from services.scan import ScanState, get_scan_service

from routes.scan import router as scan_router
from routes.protocols import router as protocols_router
from routes.pcaps import router as pcaps_router
from routes.search import router as search_router


logger = get_logger(__name__)
context = get_app_context()


# SCAN SCHEDULER
async def scheduled_scan_loop():
    """Runs in the background and triggers a scan every X seconds."""
    while True:
        try:
            interval = context.SCANNER_INTERVAL_SECONDS
            scan_service = get_scan_service()

            await asyncio.sleep(interval)

            if scan_service.scan_status["state"] == ScanState.IDLE:
                logger.info(f"Starting scheduled scan (Interval: {interval}s)...")
                loop = asyncio.get_running_loop()

                await loop.run_in_executor(
                    context.thread_executor,
                    lambda: scan_service.scan_wrapper(
                        exclude_files=None, base_url=context.FULL_BASE_URL
                    ),
                )
            else:
                logger.info("Scheduled scan skipped: Scanner is currently busy.")

        except Exception as e:
            logger.error(f"Error in scheduled scan loop: {e}")
            await asyncio.sleep(60)


app = FastAPI(
    title="Pcap Catalog Service",
    description="A service to index and search pcap files by protocol using a Redis-native inverted index.",
)

app.include_router(scan_router)
app.include_router(protocols_router)
app.include_router(pcaps_router)
app.include_router(search_router)

app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)


# --- API Endpoints ---
@app.on_event("startup")
async def startup_event():
    redis_client = context.redis_client
    await context.initialize_async()
    scan_service = get_scan_service()

    @backoff.on_exception(
        backoff.expo, redis.exceptions.ConnectionError, max_tries=5, factor=0.5
    )
    async def check_redis_for_pcaps():
        if not redis_client:
            raise redis.exceptions.ConnectionError("Redis client not initialized.")
        return await asyncio.to_thread(redis_client.keys, "pcap:file:*")

    if redis_client:
        try:
            existing_pcap_keys = await check_redis_for_pcaps()

            if not existing_pcap_keys:
                logger.info("No indexed pcaps found in Redis. Starting initial scan...")
                loop = asyncio.get_event_loop()
                loop.run_in_executor(
                    context.thread_executor,
                    lambda: scan_service.scan_wrapper(
                        exclude_files=None, base_url=context.FULL_BASE_URL
                    ),
                )
            else:
                logger.info(
                    f"Found {len(existing_pcap_keys)} indexed pcaps in Redis. Skipping initial full scan."
                )
        except Exception as e:
            logger.error(
                f"Failed to check Redis for existing pcaps during startup: {e}"
            )
    else:
        logger.error(
            "Redis client is not available. Cannot perform startup check or scan."
        )

    asyncio.create_task(scheduled_scan_loop())


@app.get("/health")
async def health_check():
    return {"status": "OK"}


if __name__ == "__main__":
    import uvicorn

    uvicorn.run(app, host="0.0.0.0", port=8000)
